pyproject.toml
[project]
name = "e3dc-cli"
dynamic = ["version"]
description = "Command-line tool to query E3/DC systems"
authors = [{ name = "Sebastian Waldvogel", email = "sebastian@waldvogels.de" }]
license = { text = "MIT" }
readme = "README.md"

requires-python = ">=3.9"
dependencies = [
    "pye3dc==0.9.2",
    "tzlocal==5.2",
    "jsonargparse==4.35.0",
    "rich-argparse==1.6.0",
    "pydantic==2.10.1",
]

[project.scripts]
e3dc-cli = "e3dc_cli.__main__:cli"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[dependency-groups]
lint = ["ruff>=0.8.3"]
pytest = ["pytest-cov>=6.0.0"]
docs = ["mkdocs>=1.6.1", "mkdocs-material>=9.5.49"]


# ---- Python PDM ------------------------------------------------------------------------------------------------------

[tool.pdm]
distribution = true

[tool.pdm.version]
source = "scm"

[tool.pdm.scripts]
lint = "ruff check ."
format = "ruff format ."
docs = "mkdocs serve"
tests = "pytest --cov-report term-missing --cov=e3dc_cli --verbose"

# ---- Ruff ------------------------------------------------------------------------------------------------------------

[tool.ruff]
include = ["**/*.py", "**/*.pyi", "**/pyproject.toml"]
line-length = 120

[tool.ruff.lint]
select = [
    "ARG",
    "ANN", # Warnings for missing type annotations
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "D",   # Missing docstrings (autodoc)
    "DOC", # pydoclint
    "E",   # pycodestyle - Error
    "F",   # Pyflakes
    "I",   # isort
    "W",   # pycodestyle - Warning
    "UP",  # pyupgrade
    #"PL", # Pylint
]
extend-select = ["DOC"]
preview = true
ignore = ["UP006", "UP035"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["PLC2701"]

[tool.ruff.lint.pydocstyle]
convention = "google"

# ---- pylint ----------------------------------------------------------------------------------------------------------

[tool.pylint]
load-plugins = ["pylint.extensions.docparams"]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.main]
ignore-paths = [".venv", ".git", "__pypackages__", "tests/_temp", "tests"]
accept-no-param-doc = "no"
accept-no-raise-doc = "no"
accept-no-yields-doc = "no"
default-docstring-type = "google"

# ---- mypy ------------------------------------------------------------------------------------------------------------

[tool.mypy]
strict = true
